refactor: reorganize utils
