+++ src/agent/modelHandlers/modelHandlerAnthropic.ts
@@ -26,7 +26,7 @@ import {
   getAllReplacementsRegex,
   cleanFileContent,
 } from '../../replacement/replacementUtils';
-import { extractAndLogScratchpad } from '../../utils/xmlUtils';
+import { extractAndLogScratchpad } from '../../utils/text/xmlUtils';
 
 // Local imports - agent components
 import { AgentConfig } from '../AgentConfig';
@@ -41,7 +41,7 @@ import { ToolState } from '../ToolState';
 import { MediaEntry } from '../mediaTypes';
 import { AgentStateRound } from '../AgentState';
 import { K_SLICE } from '../../utils/constants';
-import { objectToLogString } from '../../utils/stringUtils';
+import { objectToLogString } from '../../utils/text/stringUtils';
 import { calculateTokenPrice } from '../../utils/priceUtils';
 
 /**
