+++ src/agent/modelHandlers/modelHandlerKimi.ts
@@ -10,7 +10,7 @@ import { ToolState } from '../ToolState';
 import { MediaEntry } from '../mediaTypes';
 
 // Local imports - utilities
-import { convertContentToString } from '../../utils/messageUtils';
+import { convertContentToString } from '../../utils/text/messageUtils';
 import { K_SLICE, MESSAGE_PREVIEW_LENGTH } from '../../utils/constants';
 
 /**
