+++ src/commands/latexCommands.ts
@@ -6,7 +6,7 @@ import * as logger from '../logger/logUtils';
 
 // Local imports - utilities
 import { getRelativePath } from '../utils/workspaceFileUtils';
-import { sleep } from '../utils/timeUtils';
+import { sleep } from '../utils/helpers';
 import {
   applyReplacements,
   getAllReplacements,
