+++ src/logger/AgentLogger.ts
@@ -3,7 +3,7 @@
 
 // Local imports - log
 import * as logger from './logUtils';
-import { sleep } from '../utils/timeUtils';
+import { sleep } from '../utils/helpers';
 import { SHORT_SLEEP_MS } from '../utils/constants';
 
 /**
