+++ src/utils/helpers.ts
@@ -1,0 +1,23 @@
+// Utility helper functions
+
+/**
+ * Asynchronously wait for the specified number of milliseconds.
+ * @param ms Number of milliseconds to sleep
+ */
+export async function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+/**
+ * Generate a nonce string for Webview content security policies.
+ * @param length Length of the generated nonce
+ */
+export function generateNonce(length = 32): string {
+  const possible =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  let text = '';
+  for (let i = 0; i < length; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+  return text;
+}
