--- src/utils/nonceUtils.ts
@@ -1,10 +1,0 @@
-// Utility to generate nonce strings for Webview content security policy
-export function generateNonce(length = 32): string {
-  const possible =
-    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  let text = '';
-  for (let i = 0; i < length; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
-}
