+++ src/utils/text/messageUtils.ts
